Add github action workflow to run tests
* add section to readme for running tests

* add toc to readme

* use dedicated xcode build action

* fix compile error by renaming function in FMAPI.swift file

* split tests into dedicated test cases

* skip filter tests during CI

* add trigger for pushes to develop
