+++ FantasmoSDKTests/SDKLocationTests.swift
@@ -1,90 +1,30 @@
 //
-//  FantasmoSDKTests.swift
+//  SDKLocationTests.swift
 //  FantasmoSDKTests
 //
 //  Created by lucas kuzma on 8/4/21.
-//
+//  Modified by che fisher on 27/9/21
 
 import XCTest
 import CoreLocation
 import ARKit
 
-class FantasmoSDKTests: XCTestCase {
+class SDKLocationTests: XCTestCase {
+
+    override class func setUp() {
+        // Put setup code here that is run once (equal to mocha "before" hook)
+    }
 
     override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
+        // Put setup code here is run before each test (equal to mocha "beforeEach" hook)
     }
 
     override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
+        // Put teardown code that is run after each test case (equal to mocha "afterEach" hook)
     }
 
-    func testGeometricMedian() throws {
-        var locations: [CLLocation] = []
-        var median = CLLocation()
-        var expected = CLLocation()
-
-        locations.append(CLLocation(latitude: 0, longitude: 0))
-        median = CLLocation.geometricMedian(locations)
-        print(median.coordinate)
-
-        expected = CLLocation(latitude: 0, longitude: 0);
-        XCTAssertLessThan(median.degreeDistance(from: expected), 0.001)
-
-        locations = []
-        locations.append(CLLocation(latitude: 10, longitude: 0))
-        locations.append(CLLocation(latitude: -10, longitude: 0))
-        median = CLLocation.geometricMedian(locations)
-        print(median.coordinate)
-
-        expected = CLLocation(latitude: 0, longitude: 0);
-        XCTAssertLessThan(median.degreeDistance(from: expected), 0.001)
-
-        locations.append(CLLocation(latitude: 0, longitude: 10))
-        median = CLLocation.geometricMedian(locations)
-        print(median.coordinate)
-
-        expected = CLLocation(latitude: 0, longitude: 5.77);
-        XCTAssertLessThan(median.degreeDistance(from: expected), 0.01)
-
-        locations = []
-        locations.append(CLLocation(latitude: 10, longitude: 10))
-        locations.append(CLLocation(latitude: 20, longitude: 10))
-        locations.append(CLLocation(latitude: 10, longitude: 20))
-        locations.append(CLLocation(latitude: 20, longitude: 20))
-        median = CLLocation.geometricMedian(locations)
-        print(median.coordinate)
-
-        expected = CLLocation(latitude: 15, longitude: 15);
-        XCTAssertLessThan(median.degreeDistance(from: expected), 0.01)
-
-        locations.append(CLLocation(latitude: 15, longitude: 15))
-        median = CLLocation.geometricMedian(locations)
-        print(median.coordinate)
-
-        XCTAssertLessThan(median.degreeDistance(from: expected), 0.01)
-    }
-
-    func testGeometricMedianColinear() throws {
-        var locations: [CLLocation] = []
-        var median = CLLocation()
-        var expected = CLLocation()
-
-        locations = []
-        locations.append(CLLocation(latitude: 0, longitude: 0))
-        locations.append(CLLocation(latitude: 0, longitude: 10))
-        median = CLLocation.geometricMedian(locations)
-        print(median.coordinate)
-
-        expected = CLLocation(latitude: 0, longitude: 5);
-        XCTAssertLessThan(median.degreeDistance(from: expected), 0.01)
-
-        locations.append(CLLocation(latitude: 0, longitude: 20))
-        median = CLLocation.geometricMedian(locations)
-        print(median.coordinate)
-
-        expected = CLLocation(latitude: 0, longitude: 10);
-        XCTAssertLessThan(median.degreeDistance(from: expected), 0.01)
+    override class func tearDown() {
+        // Put teardown code that is run once (equal to mocha "after" hook)
     }
 
     func testLocationFusion() {
@@ -268,82 +208,4 @@ class FantasmoSDKTests: XCTestCase {
         locations.append(CLLocation(latitude: 10, longitude: 10))
         XCTAssertEqual(LocationFuser.confidence(locations), .high) // 5 samples
     }
-    
-    func testMovementFilter() {
-        let filter = FMMovementFilter()
-        var transform = simd_float4x4(1)
-        var pixelBuffer: CVPixelBuffer? = nil
-        CVPixelBufferCreate(kCFAllocatorDefault, 64, 64, kCVPixelFormatType_OneComponent8, nil, &pixelBuffer)
-        if let pixelBuffer = pixelBuffer {
-            var frame = MockFrame(fmCamera: MockCamera(transform: transform), capturedImage: pixelBuffer)
-            XCTAssertEqual(filter.accepts(frame), .rejected(reason: .movingTooLittle))
-            transform = simd_float4x4(1.1)
-            frame = MockFrame(fmCamera: MockCamera(transform: transform), capturedImage: pixelBuffer)
-            XCTAssertEqual(filter.accepts(frame), .accepted)
-            transform = simd_float4x4(1.099)
-            XCTAssertEqual(filter.accepts(frame), .rejected(reason: .movingTooLittle))
-        } else {
-            print ("Couldn't allocate mock pixel buffer")
-        }
-    }
-    func testCameraPitchFilter() {
-        let filter = FMCameraPitchFilter()
-        var pixelBuffer: CVPixelBuffer? = nil
-        CVPixelBufferCreate(kCFAllocatorDefault, 64, 64, kCVPixelFormatType_OneComponent8, nil, &pixelBuffer)
-        var pitch : Float = deg2rad(-90)
-        if let nonnilBuffer = pixelBuffer {
-            var frame = MockFrame(fmCamera: MockCamera(pitch: pitch), capturedImage: nonnilBuffer)
-            XCTAssertEqual(filter.accepts(frame), .rejected(reason: .pitchTooLow))
-            pitch = deg2rad(-65)
-            frame = MockFrame(fmCamera: MockCamera(pitch: pitch), capturedImage: nonnilBuffer)
-            XCTAssertEqual(filter.accepts(frame), .accepted)
-            pitch = deg2rad(0)
-            frame = MockFrame(fmCamera: MockCamera(pitch: pitch), capturedImage: nonnilBuffer)
-            XCTAssertEqual(filter.accepts(frame), .accepted)
-            pitch = deg2rad(30)
-            frame = MockFrame(fmCamera: MockCamera(pitch: pitch), capturedImage: nonnilBuffer)
-            XCTAssertEqual(filter.accepts(frame), .accepted)
-            pitch = deg2rad(60)
-            frame = MockFrame(fmCamera: MockCamera(pitch: pitch), capturedImage: nonnilBuffer)
-            XCTAssertEqual(filter.accepts(frame), .rejected(reason: .pitchTooHigh))
-       } else {
-            print ("Couldn't allocate mock pixel buffer")
-        }
-    }
-    
-    func testBlurFilter() {
-        let filter = FMBlurFilter()
-        let dayScan = UIImage(named: "dayScan", in: Bundle(for: type(of: self)), compatibleWith: nil)
-        let dayScanFrame = MockFrame(capturedImage: dayScan!.pixelBuffer()!)
-        let nightScan = UIImage(named: "nightScan", in: Bundle(for: type(of: self)), compatibleWith: nil)
-        let nightScanFrame = MockFrame(capturedImage: nightScan!.pixelBuffer()!)
-        
-        // nighttime passes twice because of no throughput
-        XCTAssertEqual(filter.accepts(nightScanFrame), .accepted)
-        XCTAssertEqual(filter.accepts(nightScanFrame), .accepted)
-        // daytime passes because it has enough variance
-        XCTAssertEqual(filter.accepts(dayScanFrame), .accepted)
-        XCTAssertEqual(filter.accepts(dayScanFrame), .accepted)
-        XCTAssertEqual(filter.accepts(dayScanFrame), .accepted)
-        XCTAssertEqual(filter.accepts(dayScanFrame), .accepted)
-        // nighttime is rejected 6 times because it is too blurry and the throughput is superior to 0.25 on the last 8 frames
-        XCTAssertEqual(filter.accepts(nightScanFrame), .rejected(reason: .imageTooBlurry))
-        XCTAssertEqual(filter.accepts(nightScanFrame), .rejected(reason: .imageTooBlurry))
-        XCTAssertEqual(filter.accepts(nightScanFrame), .rejected(reason: .imageTooBlurry))
-        XCTAssertEqual(filter.accepts(nightScanFrame), .rejected(reason: .imageTooBlurry))
-        XCTAssertEqual(filter.accepts(nightScanFrame), .rejected(reason: .imageTooBlurry))
-        XCTAssertEqual(filter.accepts(nightScanFrame), .rejected(reason: .imageTooBlurry))
-        // on the 7th nighttime picture in a row, throughput gets back under 0.25 and it passes again
-        XCTAssertEqual(filter.accepts(nightScanFrame), .accepted)
-        XCTAssertEqual(filter.accepts(nightScanFrame), .accepted)
-        XCTAssertEqual(filter.accepts(nightScanFrame), .accepted)
-        XCTAssertEqual(filter.accepts(nightScanFrame), .accepted)
-    }
-    //    func testPerformanceExample() throws {
-    //        // This is an example of a performance test case.
-    //        measure {
-    //            // Put the code you want to measure the time of here.
-    //        }
-    //    }
-
 }
